+++ src/fourty_eight.rs
@@ -4,9 +4,9 @@ use crate::data_item::{
     AircraftAddress, AircraftIdentification, CalculatedPositionCartesianCorr,
     CalculatedTrackVelocity, CommunicationsCapabilityFlightStatus, DataSourceIdentifier,
     FlightLevelInBinaryRepresentation, MeasuredPositionInPolarCoordinates,
-    Mode3ACodeConfidenceIndicator, Mode3ACodeInOctalRepresentation, ModeSMBData,
-    RadarPlotCharacteristics, TargetReportDescriptor, TimeOfDay, TrackNumber, TrackQuality,
-    TrackStatus, WarningErrorConditionsTargetClass,
+    Mode3ACodeConfidenceIndicator, Mode3ACodeInOctalRepresentation,
+    ModeCCodeAndConfidenceIndicator, ModeSMBData, RadarPlotCharacteristics, TargetReportDescriptor,
+    TimeOfDay, TrackNumber, TrackQuality, TrackStatus, WarningErrorConditionsTargetClass,
 };
 use crate::fspec::{add_fx, is_fspec, read_fspec, trim_fspec};
 
@@ -68,8 +68,11 @@ pub struct Cat48 {
         cond = "is_fspec(Mode3ACodeConfidenceIndicator::FRN_48, fspec, 2)"
     )]
     pub mode3a_code_confidence_indicator: Option<Mode3ACodeConfidenceIndicator>,
-    //#[deku(skip, cond = "is_fspec(0b1_0000, fspec, 2)")]
-    //pub modec_code_and_confidence_indicator = Option<ModeCCodeAndConfidenceIndicator>,
+    #[deku(
+        skip,
+        cond = "is_fspec(ModeCCodeAndConfidenceIndicator::FRN_48, fspec, 2)"
+    )]
+    pub modec_code_and_confidence_indicator: Option<ModeCCodeAndConfidenceIndicator>,
     //#[deku(skip, cond = "is_fspec(0b1000, fspec, 2)")]
     //pub height_measured_by_3d_radar = Option<HeightMeasuredBy3dRadar>,
     //#[deku(skip, cond = "is_fspec(0b100, fspec, 2)")]
@@ -137,6 +140,9 @@ impl Cat48 {
         }
         if self.mode3a_code_confidence_indicator.is_some() {
             fspec[2] |= Mode3ACodeConfidenceIndicator::FRN_48;
+        }
+        if self.modec_code_and_confidence_indicator.is_some() {
+            fspec[2] |= ModeCCodeAndConfidenceIndicator::FRN_48;
         }
         if self.communications_capability_flight_status.is_some() {
             fspec[2] |= CommunicationsCapabilityFlightStatus::FRN_48;